refactor(dashboard): decompose trpg.ts and ops.ts into subdirectories (P1)
Split two god modules into focused files following the established
components/live/ pattern:

- trpg.ts (1545 lines) → trpg/ (5 files: helpers, index, views,
  sub-components, controls)
- ops.ts (1127 lines) → ops/ (2 files: helpers, index)

Barrel index.ts files preserve external import paths unchanged.
Pure structural refactoring — no behavior changes.
